+++ src/constants/include/constants/CGlobal.hpp
@@ -11,11 +11,17 @@
  */
 struct CGlobal final {
 	// config counts
-	static inline int const configEntryCount{ 4 }; ///< config load checks if the count is idetical to the loaded entry count
-	static inline int const configVersionCount{ 2 }; ///< config load checks if the count is idetical to the loaded version count
-	static inline int const configSectionCount{ 10 }; ///< config load checks if the count is idetical to the loaded section count
+    static inline size_t const configEntryCount{
+        4
+    }; ///< config load checks if the count is idetical to the loaded entry count
+    static inline size_t const configVersionCount{
+        2
+    }; ///< config load checks if the count is idetical to the loaded version count
+    static inline size_t const configSectionCount{
+        10
+    }; ///< config load checks if the count is idetical to the loaded section count
 
-	// versions
+    // versions
 	static inline std::string const languageVersion{ "v0.1.5.4" }; ///< contains the language version string
 	static inline std::string const configVersion{ "v0.1.5.0" }; ///< contains the config version string
 	// gameVersion will be extracted with regex in the GitHub CI. Don't chance the following line. see CMakeList.txt
